+++ hw3/src/com/company/ThingForSale.java
@@ -1,8 +1,8 @@
 package com.company;
 
 public interface ThingForSale {
-    void setPrice(int price);
+    void setPrice(double price);
     void setName(String name);
     void setCategory(String category);
-    int getPrice();
+    double getPrice();
 }
